pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift-watchdog"
version = "1.0.0"
description = "Lightweight ML model drift detection — CLI, Prometheus metrics, and alerts"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "drift-watchdog contributors"}
]
keywords = ["ml", "drift-detection", "monitoring", "prometheus", "mlops"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.14.0",
    "requests>=2.28.0",
    "boto3>=1.26.0",
    "google-cloud-storage>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.990",
    "flake8>=5.0.0",
]

[project.scripts]
drift-watchdog = "drift_watchdog.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/drift-watchdog"
Repository = "https://github.com/your-org/drift-watchdog"
Issues = "https://github.com/your-org/drift-watchdog/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
